1.修改基建物资下预订单时，可编辑带小数的数量订单 2.完善基建物资周转材料/设备预定订单功能逻辑 3.新增基建物资详情周转材料/设备可与商家聊天 4.适配android10本地图片上传与相机拍照，已调试成功华为机型

+++ app/src/main/java/com/shuangduan/zcy/model/api/ApiService.java
@@ -1538,7 +1538,7 @@ public interface ApiService {
             @Field("user_id") int userId,
             @Field("id") int id,
             @Field("remark") String remark,
-            @Field("buy_stock") long buyStock,
+            @Field("buy_stock") double buyStock,
             @Field("lease_start_time") String leaseStartTime,
             @Field("lease_end_time") String leaseEndTime,
             @Field("address_id") int address_id
